fix: UI and Search improvements
This address visiblity issues in the UI by placing an outline around the profile-photo of the user in order to increase visiblity during loading and error states.  Inconssitencies in search result state are also fixed whereby some states might not be present leaving a trailing comma in the search result text.

Overall, this leads to a smoother user experience for users